% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGeneric.R
\name{content<-}
\alias{content<-}
\title{Set content type}
\usage{
content(object) <- value
}
\arguments{
\item{object}{a \code{DrugSyn} object}

\item{value}{"death" or "survival"}
}
\description{
Set content type
}

