% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{detectDec}
\alias{detectDec}
\title{Automatic detection of decimal character in file}
\usage{
detectDec(file)
}
\arguments{
\item{file}{input file}
}
\value{
character "." or ","
}
\description{
Automatic detection of decimal character in file
}

