% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGeneric.R
\name{drugNames<-}
\alias{drugNames<-}
\title{Set drug names}
\usage{
drugNames(object) <- value
}
\arguments{
\item{object}{a \code{DrugSyn} object}

\item{value}{character vector}
}
\description{
Set drug names
}

