% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synDrug.R
\docType{class}
\name{DrugSyn-class}
\alias{DrugSyn-class}
\title{Title}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{data}}{data.frame.}

\item{\code{doses}}{list.}

\item{\code{respInd}}{list.}

\item{\code{drugNames}}{vector.}

\item{\code{dataMean}}{data.frame.}

\item{\code{content}}{character.}

\item{\code{typeHill}}{numeric.}

\item{\code{range}}{character.}
}}

