% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGeneric.R
\name{expData<-}
\alias{expData<-}
\title{Set experimental data}
\usage{
expData(object) <- value
}
\arguments{
\item{object}{a \code{DrugSyn} object}

\item{value}{a \code{data.frame}}
}
\description{
Set experimental data
}

