% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hill.R
\name{hill}
\alias{hill}
\title{Hill function}
\usage{
hill(x, Emax, E0, n, EC50)
}
\arguments{
\item{x}{dose to be evaluated}

\item{Emax}{Emax parameter of hill function}

\item{E0}{\eqn{E_0} parameter of hill function}

\item{n}{slope parameter of hill function}

\item{EC50}{EC50 parameter of hill function}
}
\value{
numeric
}
\description{
Hill function
}
\examples{
hill(0.5, 100, 0, 2, 0.5)
}

