% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{isobologram}
\alias{isobologram}
\title{Title}
\usage{
isobologram(object, effect = 0.5, mode = c("linear", "SP"))
}
\arguments{
\item{object}{a DrugSyn object}

\item{effect}{numeric, effect at which the isobologram is plotted}

\item{mode}{type of the isobologram, see Details}
}
\description{
Title
}
\details{

}

