% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loewe.R
\name{loewe}
\alias{loewe}
\title{Compute the predicted effect under Loewe model}
\usage{
loewe(a, b, modA, modB)
}
\arguments{
\item{a}{dose of drug A}

\item{b}{dose of drug B}

\item{modA}{hill model for drug A}

\item{modB}{hill model for drug B}
}
\value{
expected effect of combination under Loewe additive model
}
\description{
Compute the predicted effect under Loewe model
}
\examples{
NULL
}
