% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synDrug.R
\name{makeDrugSyn}
\alias{makeDrugSyn}
\title{Create DrugSyn object}
\usage{
makeDrugSyn(data, doses, content = c("Death", "Survival"), typeHill = c(4,
  3), range = c("Fraction", "Percentage"))
}
\arguments{
\item{range}{}
}
\value{
a \code{DrugSyn} object
}
\description{
Create DrugSyn object
}

