% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mySeq}
\alias{mySeq}
\title{Generate sequence of doses}
\usage{
mySeq(doses, n)
}
\arguments{
\item{doses}{data doses}

\item{n}{number of points in each doses interval (inclusive)}
}
\value{
sequence of doses
}
\description{
Generate sequence of doses
}

