% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{parPlot}
\alias{parPlot}
\title{Plot DrugSyn object as parallel lines}
\usage{
parPlot(object, ref, xaxis = c("pos", "dose"))
}
\arguments{
\item{object}{a \code{DrugSyn} object}

\item{ref}{character, name of the drug plotted as x-axis}

\item{xaxis}{either "pos" or "dose", indicates if the scale of x-axis is in doses of positions in data matrix}
}
\value{
a \code{ggplot} object
}
\description{
Plot DrugSyn object as parallel lines
}
\examples{
data(exDrugSyn)
parPlot(exDrugSyn, ref = "drugA")
}
