% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synDrug.R
\docType{methods}
\name{plot,DrugSyn,ANY-method}
\alias{plot,DrugSyn,ANY-method}
\title{Plot DrugSyn object}
\usage{
\S4method{plot}{DrugSyn,ANY}(x, type = c("heatmap", "parallel", "ind",
  "surface"), ...)
}
\arguments{
\item{x}{DrugSyn object}

\item{type}{type of plot}

\item{...}{arguments given to plots functions}
}
\description{
Plot DrugSyn object
}

