% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot drugSyn object as heatmap}
\usage{
plotHeatmap(object, what = c("value", "HSA", "Chou", "Bliss", "Loewe",
  "LoeweExcess"), addVal = TRUE, main = "")
}
\arguments{
\item{object}{a DrugSyn object}

\item{what}{character containing the name of the variable to plot}

\item{addVal}{logical indicating if values should be displayed on heatmap}

\item{main}{character containing title of the plot}
}
\value{
a ggplot object
}
\description{
Plot drugSyn object as heatmap
}
\examples{
data(exDrugSyn)
## Experimental data
plotHeatmap(exDrugSyn)
## Loewe predicted effect
plotHeatmap(exDrugSyn, what = "Loewe")
## Chou CI
plotHeatmap(exDrugSyn, what = "Chou")
}
