% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotSurface}
\alias{plotSurface}
\title{Title}
\usage{
plotSurface(object, what = c("value", "Loewe", "LoeweExcess", "HSA", "Bliss",
  "Chou"), gridDensity = 4, ...)
}
\arguments{
\item{object}{a drugSyn object}

\item{what}{data to plot}

\item{gridDensity}{density of the grid for interpolation}

\item{...}{arguments to persp}
}
\value{
a plot
}
\description{
Title
}

