% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGeneric.R
\name{respInd<-}
\alias{respInd<-}
\title{Set individual responses}
\usage{
respInd(object) <- value
}
\arguments{
\item{object}{a \code{DrugSyn} object}

\item{value}{list}
}
\description{
Set individual responses
}
