% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{respPlot}
\alias{respPlot}
\title{Plot individual drug response}
\usage{
respPlot(object, drug, xlog = TRUE)
}
\arguments{
\item{object}{a \code{DrugSyn} object}

\item{drug}{character, name of the drug}

\item{xlog}{boolean, describe if th x-axis is in log-scale}
}
\value{
a \code{ggplot} object
}
\description{
Plot individual drug response
}
\examples{
data(exDrugSyn)
respPlot(exDrugSyn, "drugA")
}

