% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hill.R
\name{revhill}
\alias{revhill}
\title{Inverse function of Hill equation}
\usage{
revhill(E, Emax, E0, n, EC50)
}
\arguments{
\item{E}{effect}

\item{Emax}{Emax parameter of hill function}

\item{E0}{\eqn{E_0} parameter of hill function}

\item{n}{slope parameter of hill function}

\item{EC50}{EC50 parameter of hill function}
}
\value{
numeric
}
\description{
Inverse function of Hill equation
}
\examples{
revhill(50, 100, 0, 2, 0.5)
}

