% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp.R
\name{runSynerDrug}
\alias{runSynerDrug}
\title{Run synerdrug shiny interface}
\usage{
runSynerDrug()
}
\value{
Opens web browser on shiny interface.
}
\description{
Run synerdrug shiny interface
}
\examples{
\dontrun{
runSynerDrug()
}
}
