% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{synergy}
\alias{synergy}
\title{Experimental drug synergy data}
\format{a data.frame with 4 columns:
\itemize{
 \item value fraction of killed cells
 \item rep replicate id
 \item drugA concentration of drug A
 \item drugB concentration of drug B
}}
\usage{
synergy
}
\description{
Experimental drug synergy data
}
\keyword{datasets}

