% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synDrug.R
\docType{class}
\name{DrugSyn-class}
\alias{DrugSyn-class}
\alias{content,DrugSyn-method}
\alias{content<-,DrugSyn,character-method}
\alias{doses,DrugSyn-method}
\alias{doses<-,DrugSyn,list-method}
\alias{drugNames,DrugSyn-method}
\alias{drugNames<-,DrugSyn,vector-method}
\alias{expData,DrugSyn-method}
\alias{expData<-,DrugSyn,data.frame-method}
\alias{plot,DrugSyn-method}
\alias{respInd,DrugSyn-method}
\alias{respInd<-,DrugSyn,list-method}
\title{DrugSyn class}
\usage{
\S4method{doses}{DrugSyn}(object)

\S4method{doses}{DrugSyn,list}(object) <- value

\S4method{expData}{DrugSyn}(object)

\S4method{expData}{DrugSyn,data.frame}(object) <- value

\S4method{respInd}{DrugSyn}(object)

\S4method{respInd}{DrugSyn,list}(object) <- value

\S4method{drugNames}{DrugSyn}(object)

\S4method{drugNames}{DrugSyn,vector}(object) <- value

\S4method{content}{DrugSyn}(object)

\S4method{content}{DrugSyn,character}(object) <- value

\S4method{plot}{DrugSyn}(x, type = c("heatmap", "parallel", "ind", "surface"),
  ...)
}
\arguments{
\item{object}{a DrugSyn}

\item{value}{replacement value}

\item{x}{DrugSyn object}

\item{type}{type of plot}

\item{...}{arguments given to plots functions}
}
\description{
S4 class to store drug combination data
}
\section{Methods (by generic)}{
\itemize{
\item \code{doses}: get doses values

\item \code{doses<-}: set doses values

\item \code{expData}: extract experimental data

\item \code{expData<-}: set experimental data

\item \code{respInd}: extract individual responses

\item \code{respInd<-}: set individual responses

\item \code{drugNames}: extract drug names

\item \code{drugNames<-}: set drug names

\item \code{content}: get content type

\item \code{content<-}: set content type

\item \code{plot}: plot DrugSyn object
}}
\section{Slots}{

\describe{
\item{\code{data}}{data.frame}

\item{\code{doses}}{list}

\item{\code{respInd}}{list}

\item{\code{drugNames}}{vector}

\item{\code{dataMean}}{data.frame}

\item{\code{content}}{character}

\item{\code{typeHill}}{numeric}

\item{\code{range}}{character}
}}

