% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hill.R
\name{hill}
\alias{hill}
\title{Hill function}
\usage{
hill(x, Emax, E0, n, EC50)
}
\arguments{
\item{x}{}

\item{Emax}{}

\item{E0}{}

\item{n}{}

\item{EC50}{}
}
\value{
numeric
}
\description{
Hill function
}
\examples{
hill(0.5, 100, 0, 2, 0.5)
}

