% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synDrug.R
\docType{methods}
\name{plot,DrugSyn,ANY-method}
\alias{plot,DrugSyn,ANY-method}
\title{Title}
\usage{
\S4method{plot}{DrugSyn,ANY}(x, type = c("heatmap", "parallel", "ind",
  "surface"), ...)
}
\arguments{
\item{x}{DrugSyn.}
}
\description{
Title
}

