% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotSurface}
\alias{plotSurface}
\title{Plot drugSyn object as surface}
\usage{
plotSurface(object, what = c("value", "Loewe", "LoeweExcess", "HSA", "Bliss",
  "Chou"), gridDensity = 4, ...)
}
\arguments{
\item{object}{a \code{DrugSyn} object}

\item{what}{data to plot}

\item{gridDensity}{density of the grid for interpolation}

\item{...}{other arguments passed to \code{\link{persp}}}
}
\value{
a plot
}
\description{
Plot drugSyn object as surface
}
\examples{
#' data(exDrugSyn)
## Experimental data
plotSurface(exDrugSyn)
## change azimuth
plotSurface(exDrugSyn, theta = -80)
## Loewe excess
plotSurface(exDrugSyn, what = "LoeweExcess")
}

